feat: グループIDを扱いやすくするため、サーチパラメータに変更、findGroupById関数を別ファイルに分離 #114
